Slight improvements. Tweets table added and now we can obtain tweet text if necessary. Also hasTweets and mentions relationships implemented in BFS and DFS algorithms

+++ TwitterCrawler/src/main/java/dto/UserDto.java
@@ -19,7 +19,7 @@ import javax.persistence.Table;
 @Table(name="Users")
 public class UserDto extends NodeDto {
 
-	private static int k=0;
+	private static long k=0;
 	
 	public UserDto(){
 		++k;
@@ -79,7 +79,7 @@ public class UserDto extends NodeDto {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
